Add: Money type and refactor related method in VendingMachine
